move courses files to imports
