+++ client/main.scss
@@ -32,21 +32,9 @@
 @import '{}/imports/ui/stylesheets/rtl';
 
 // Template styles
-@import '{}/imports/ui/components/categories/categories';
-@import '{}/imports/ui/components/editable/editable';
-@import '{}/imports/ui/components/footer/footer';
-@import '{}/imports/ui/components/introduction/introduction';
-@import '{}/imports/ui/components/loading/loading';
-@import '{}/imports/ui/components/map/map';
-@import '{}/imports/ui/components/messages/messages';
-@import '{}/imports/ui/components/navbar/navbar';
-@import '{}/imports/ui/components/price-policy/price-policy';
-@import '{}/imports/ui/components/region-tag/region-tag';
-@import '{}/imports/ui/components/report/report';
-@import '{}/imports/ui/components/sharing/sharing';
-@import '{}/imports/ui/components/translate-info/translate-info';
+@import '{}/imports/ui/components/index';
 
-@import '{}/imports/ui/pages/index.scss';
+@import '{}/imports/ui/pages/index';
 
 // Stylesheets from old structure outside of imports
 @import 'views/views';
