+++ client/views/_views.scss
@@ -1,17 +1,6 @@
 // Views stylesheet imports
 // --------------------------------------------------
 
-// Courses
-@import 'courses/compact/course.compact';
-@import 'courses/details/course.details';
-@import 'courses/discussion/course.discussion';
-@import 'courses/history/course.history';
-@import 'courses/members/course.members';
-@import 'courses/roles/course.roles';
-@import 'courses/edit/course.edit';
-@import 'courses/events/course.events';
-@import 'courses/list/course.list';
-
 // Events
 @import 'events/compact/event.compact';
 @import 'events/details/event.details';
