+++ imports/ui/components/courses/course-compact/course-compact.js
@@ -1,4 +1,11 @@
+import { Meteor } from 'meteor/meteor';
+import { Template } from 'meteor/templating';
+import { $ } from 'meteor/jquery';
+import { _ } from 'meteor/underscore';
+
 import '/imports/ui/components/categories/categories.js';
+
+import './course-compact.html';
 
 Template.courseCompact.helpers({
 	ready: function() {
