+++ imports/ui/components/courses/course-compact/course-compact.scss
