+++ imports/ui/components/courses/course-discussion/course-discussion.js
@@ -1,6 +1,13 @@
+import { Meteor } from 'meteor/meteor';
+import { ReactiveVar } from 'meteor/reactive-var';
+import { Template } from 'meteor/templating';
+import { $ } from 'meteor/jquery';
+
 import { AddMessage } from '/imports/api/messages/methods.js';
 
 import '/imports/ui/components/buttons/buttons.js';
+
+import './course-discussion.html';
 
 Template.discussion.onCreated(function() {
 	this.count = new ReactiveVar(0);
