+++ imports/ui/components/courses/course-events/course-events.html
