+++ imports/ui/components/courses/course-history/course-history.js
@@ -1,0 +1,12 @@
+import { Template } from 'meteor/templating';
+
+import './course-history.html';
+
+Template.coursehistory.helpers({
+	pastEventsList: function() {
+		return Events.find(
+			{ courseId: this.course._id, start: { $lt: new Date() } },
+			{ sort: { start: -1 } }
+		);
+	},
+});
