+++ imports/ui/components/courses/course-list/course-list.html
