+++ imports/ui/components/courses/course-list/course-list.js
@@ -1,0 +1,3 @@
+import '../course-compact/course-compact.js';
+
+import './course-list.html';
