+++ imports/ui/components/index.scss
@@ -1,0 +1,21 @@
+@import './categories/categories';
+@import './courses/course-compact/course-compact';
+@import './courses/course-discussion/course-discussion';
+@import './courses/course-edit/course-edit';
+@import './courses/course-events/course-events';
+@import './courses/course-history/course-history';
+@import './courses/course-list/course-list';
+@import './courses/course-members/course-members';
+@import './courses/course-roles/course-roles';
+@import './editable/editable';
+@import './footer/footer';
+@import './introduction/introduction';
+@import './loading/loading';
+@import './map/map';
+@import './messages/messages';
+@import './navbar/navbar';
+@import './price-policy/price-policy';
+@import './region-tag/region-tag';
+@import './report/report';
+@import './sharing/sharing';
+@import './translate-info/translate-info';
