+++ imports/ui/lib/course-template.js
@@ -1,0 +1,6 @@
+export default function CourseTemplate() {
+	return {
+		roles: ['host', 'mentor'],
+		region: Session.get('region')
+	};
+};
