+++ imports/ui/pages/course-create/course-create.js
@@ -1,0 +1,1 @@
+import './course-create.html';
