+++ imports/ui/pages/course-details/course-details.html
