+++ imports/ui/pages/course-details/course-details.scss
