+++ imports/ui/pages/frames/propose-frame/propose-frame.js
@@ -1,1 +1,3 @@
+import '/imports/ui/components/courses/course-edit/course-edit.js';
+
 import './propose-frame.html';
