+++ imports/ui/pages/index.scss
@@ -1,4 +1,5 @@
 @import './calendar/calendar';
+@import './course-details/course-details';
 @import './faq/faq';
 @import './frames/events-frame/events-frame';
 @import './frames/schedule-frame/schedule-frame';
